// generated/models/ApiMarketBatchPrepareRequest.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiMarketBatchItemRequest } from '../models/ApiMarketBatchItemRequest';
import { HttpFile } from '../http/http';

export class ApiMarketBatchPrepareRequest {
    'kind': ApiMarketBatchPrepareRequestKindEnum;
    'profile_id': string;
    'wallet': string;
    'currency': ApiMarketBatchPrepareRequestCurrencyEnum;
    'execution_policy': ApiMarketBatchPrepareRequestExecutionPolicyEnum;
    'amount_wei': string;
    'items': Array<ApiMarketBatchItemRequest>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "kind",
            "baseName": "kind",
            "type": "ApiMarketBatchPrepareRequestKindEnum",
            "format": ""
        },
        {
            "name": "profile_id",
            "baseName": "profile_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "wallet",
            "baseName": "wallet",
            "type": "string",
            "format": ""
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "ApiMarketBatchPrepareRequestCurrencyEnum",
            "format": ""
        },
        {
            "name": "execution_policy",
            "baseName": "execution_policy",
            "type": "ApiMarketBatchPrepareRequestExecutionPolicyEnum",
            "format": ""
        },
        {
            "name": "amount_wei",
            "baseName": "amount_wei",
            "type": "string",
            "format": ""
        },
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<ApiMarketBatchItemRequest>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiMarketBatchPrepareRequest.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiMarketBatchPrepareRequestKindEnum {
    BuyBatch = 'BUY_BATCH'
}
export enum ApiMarketBatchPrepareRequestCurrencyEnum {
    _0x0000000000000000000000000000000000000000 = '0x0000000000000000000000000000000000000000'
}
export enum ApiMarketBatchPrepareRequestExecutionPolicyEnum {
    AllOrRevert = 'ALL_OR_REVERT'
}